teaching_ppt/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-pptx>=0.6.23
